7 TelegramLongPollingBot для получения прогресса

+++ src/main/java/org/mozgotrash/rest/response/ProgressDto.java
@@ -5,9 +5,11 @@ import lombok.Data;
 import org.mozgotrash.model.Book;
 import org.mozgotrash.model.Goal;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ProgressDto {
     GoalDto goal;
-    double progressPercentage;
+    BigDecimal progressPercentage;
 }
